// types/ui/ui.types.ts
export type ViewType = 'menu'
  | 'game'
  | 'results'
  | 'archive'
  | 'howToPlay'
  | 'contact'
  | 'userStats'
  | 'reminders'
  | 'unsubscribe'
  | 'privacy'
  | 'terms'

export type ModalType = 'emailReminder'
  | 'share'
  | 'rateLimit'
  | 'sharedPuzzleNotFound'
  | 'serviceUnavailable'
  | 'options'
  | 'globalStats'
  | 'gameTutorial'
  | null